Use the locally checked out sources to check
`pod spec lint` will use the value of the `spec.source` attribute,
which points to the latest release in our repo. `pod spec lib` uses
the locally checked out sources.

I'll probably add a new build job which uses `pod spec lint` which
will be used by the CI when we release in a subsequent PR.